ADD THE SCRIPT HELPER TO THE HELPERS MENU
Previously the only way to invoke this helper was by specifying it on
the command line not by selecting it from the menu. This also meant that
the script name had to be specified on the command line in the same
command. Now you can select this helper from the menu and then add the
script name later, when prompted for it.